refactor: moved productService to product package
